Simplify some more type resolutions
