+++ src/components/views/dialogs/CreateSpaceFromCommunityDialog.tsx
@@ -43,56 +43,13 @@ import TagOrderActions from "../../../actions/TagOrderActions";
 import { inviteUsersToRoom } from "../../../RoomInvite";
 import ProgressBar from "../elements/ProgressBar";
 import { ViewRoomPayload } from "../../../dispatcher/payloads/ViewRoomPayload";
+import { CreateEventField, IGroupRoom, IGroupSummary } from "../../../@types/groups";
 
 interface IProps {
     matrixClient: MatrixClient;
     groupId: string;
     onFinished(spaceId?: string): void;
 }
-
-export const CreateEventField = "io.element.migrated_from_community";
-
-interface IGroupRoom {
-    displayname: string;
-    name?: string;
-    roomId: string;
-    canonicalAlias?: string;
-    avatarUrl?: string;
-    topic?: string;
-    numJoinedMembers?: number;
-    worldReadable?: boolean;
-    guestCanJoin?: boolean;
-    isPublic?: boolean;
-}
-
-/* eslint-disable camelcase */
-export interface IGroupSummary {
-    profile: {
-        avatar_url?: string;
-        is_openly_joinable?: boolean;
-        is_public?: boolean;
-        long_description: string;
-        name: string;
-        short_description: string;
-    };
-    rooms_section: {
-        rooms: unknown[];
-        categories: Record<string, unknown>;
-        total_room_count_estimate: number;
-    };
-    user: {
-        is_privileged: boolean;
-        is_public: boolean;
-        is_publicised: boolean;
-        membership: string;
-    };
-    users_section: {
-        users: unknown[];
-        roles: Record<string, unknown>;
-        total_user_count_estimate: number;
-    };
-}
-/* eslint-enable camelcase */
 
 enum Progress {
     NotStarted,
